requirements.txt
python-telegram-bot
flask
git+https://github.com/yt-dlp/yt-dlp.git